requirements.txt

Flask>=2.2.2
gunicorn==20.0.4
python-telegram-bot==13.1
python-youtube==0.6.4
requests==2.25.1
wikipedia==1.4.0
Wikipedia-API==0.5.4
youtube-dl>=2021.2.4.1
youtube-search-python>=1.3.4
